refactor: separate upgrades into module
