+++ crates/asm/subprotocols/upgrade/src/subprotocol.rs
@@ -2,12 +2,14 @@ use strata_asm_common::{MsgRelayer, NullMsg, Subprotocol, SubprotocolId, TxInput
 
 use crate::{
     actions::{
-        UpgradeAction,
         cancel::{CANCEL_TX_TYPE, handle_cancel_tx},
-        multisig_update::{MULTISIG_CONFIG_UPDATE_TX_TYPE, handle_multisig_config_update_tx},
-        operator_update::{OPERATOR_UPDATE_TX_TYPE, handle_operator_update_tx},
-        seq_update::{SEQUENCER_UPDATE_TX_TYPE, handle_sequencer_update_tx},
-        vk_update::{VK_UPDATE_TX_TYPE, handle_vk_update_tx},
+        upgrades::{
+            UpgradeAction,
+            multisig::{MULTISIG_CONFIG_UPDATE_TX_TYPE, handle_multisig_config_update_tx},
+            operator::{OPERATOR_UPDATE_TX_TYPE, handle_operator_update_tx},
+            seq::{SEQUENCER_UPDATE_TX_TYPE, handle_sequencer_update_tx},
+            vk::{VK_UPDATE_TX_TYPE, handle_vk_update_tx},
+        },
     },
     roles::StrataProof,
     state::UpgradeSubprotoState,
